requirements.txt
fastapi
redis
types-redis
uvicorn
mysql-connector-python
python-dotenv
